Production security values.
On production servers it is nice to expose as little data as possible about the
platform and versions running things. Bot-style attacks often look for certain
fingerprints to decide to continue an attack like a vulerability on a specific
Ubuntu or PHP version. On production sites we can limit this information.

  * Do not show indexes for directories.
  * Hide Apache version number.
  * Do not show OS information in Apache responses.
  * Hide PHP information in responses.